Tortoise_ORM/exam1/requirements.txt
aiohttp==3.6.2
aiomysql==0.0.20
aiosqlite==0.12.0
alembic==1.4.2
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
bcrypt==3.1.7
cffi==1.14.0
chardet==3.0.4
ciso8601==2.1.3
click==7.1.1
colorama==0.4.3
cryptography==2.9
databases==0.2.6
fastapi==0.54.1
gino==1.0.0
h11==0.9.0
hiredis==1.0.1
httptools==0.1.1
idna==2.9
iso8601==0.1.12
jwt==1.0.0
loguru==0.4.1
Mako==1.1.2
MarkupSafe==1.1.1
multidict==4.7.5
mysql==0.0.2
mysql-connector-python==8.0.20
mysqlclient==1.4.6
passlib==1.7.2
protobuf==3.11.3
pycparser==2.20
pydantic==1.5.1
PyJWT==1.6.4
PyMySQL==0.9.3
PyPika==0.37.4
python-dateutil==2.8.1
python-decouple==3.1
python-editor==1.0.4
python-multipart==0.0.5
six==1.14.0
SQLAlchemy==1.3.16
starlette==0.13.2
tortoise-orm==0.16.10
typing-extensions==3.7.4.2
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
yarl==1.4.2